Finslip

+++ PHP-Forum/src/DeleteTopicController.php
@@ -15,17 +15,43 @@
 			$this->model = new LoginModel($userAgent);
 			$this->ShowForumView = new ForumView($this->model);
 			$this->db = new DBDetails();
+			$this->doControll();
 
 		}
 		//anropar vilken vy som ska visas.
+		public function doControll()
+		{
+			try{
+				if($this->ShowForumView->didUserPressDeleteTopic() && $this->model->checkLoginStatus() && $this->model->getLoggedInUserRole() == 1){
+					
+					//Tar bort en specifik topic med hjälp av ett id
+					$this->db->DeleteTopic($this->ShowForumView->getTopicId(), $this->model->getLoggedInUser());
+					$this->ShowForumView->successfulDeletedTopic();
+					
+	
+					
+				}
+				else
+				{
+					$this->db->LogAction($this->model->getLoggedInUser(), "Failed to delete topic with id " . $this->ShowForumView->getTopicId());
+					throw new Exception("You don't have rights to do this");
+				}
+			}
+			catch(Exception $e)
+			{
+					
+				$this->ShowForumView->showMessage($e->getMessage());
+			}
+			$this->doHTMLBody();
+
+				
+				
+			
+
+		}
 		public function doHTMLBody()
 		{
 			if($this->ShowForumView->didUserPressDeleteTopic() && $this->model->checkLoginStatus()){
-
-				//Tar bort en specifik topic med hjälp av ett id
-				$this->db->DeleteTopic($this->ShowForumView->getTopicId(), $this->model->getLoggedInUser());
-				
-
 				//När topic har tagits bort så visas alla topics
 				$topics = $this->db->fetchAllTopics();
 				$this->ShowForumView->ShowAllTopics($topics);
